+++ redirect/tasks.py
@@ -1,8 +1,7 @@
 import datetime
 
-from django.db import connection
-
 from redirect.models import Redirect, RedirectArchive
+from universal.helpers import make_chunks
 
 
 def expired_to_archive_table():
@@ -69,15 +68,3 @@ def copy_redirect(model, existing_redirect):
                  job_location=existing_redirect.job_location,
                  job_title=existing_redirect.job_title,
                  company_name=existing_redirect.company_name)
-
-
-def make_chunks(l, n=1025):
-    """
-    Yield successive n-sized chunks from a list.
-
-    """
-    if connection.vendor == 'sqlite':
-        # SQLite has a default maximum number of SQL variables of 999
-        n = min(n, 999)
-    for i in xrange(0, len(l), n):
-        yield l[i:i+n]
